Function window_vibrancy::apply_acrylic

pub fn apply_acrylic(
    window: impl HasWindowHandle,
    color: Option<Color>,
) -> Result<(), Error>

Applies acrylic effect to window. Works only on Windows 10 v1809 or newer.

WARNING:

This method has poor performance on Windows 10 v1903+ and Windows 11 build 22000, the window will lag when resizing or dragging. It is an issue in the undocumented api used for this method and microsoft needs to fix it (they probably won’t).

Platform-specific

Windows: color is ignored on Windows 7 and has no effect.

Linux / macOS: Unsupported.